requirements.txt
# FastAPI ------------------------------------------------------------------------
fastapi>=0.86.0
uvicorn>=0.19.0
python-multipart>=0.0.5

# Base ------------------------------------------------------------------------
ipython>=8.6.0 
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python--headless>=4.1.1
Pillow>=7.1.2
psutil>=5.9.4  
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
thop>=0.1.1  
torch>=1.7.0  
torchvision>=0.8.1
tqdm>=4.64.0
requests>=2.28.1

# Logging ---------------------------------------------------------------------
tensorboard>=2.4.1

# Plotting --------------------------------------------------------------------
pandas>=1.1.4
seaborn>=0.11.0